pyproject.toml
[project]
name = "boss-sentinel"
version = "1.1.0"
description = "基于YOLOv8的Windows哨兵系统 - 人脸识别与自动锁屏"
readme = "README.md"
requires-python = ">=3.9,<3.13"
authors = [
    { name = "Boss Sentinel Team" }
]
keywords = ["face-recognition", "yolo", "monitoring", "security", "windows"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Security",
    "Topic :: Multimedia :: Video :: Capture",
]

dependencies = [
    "opencv-python>=4.5.0",
    "ultralytics>=8.0.0",
    "facenet-pytorch>=2.5.0",
    "numpy>=1.21.0",
    "schedule>=1.1.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pywin32>=300",
    "Pillow>=9.0.0",
    "PyQt5>=5.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
boss-sentinel = "boss_sentinel.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
